requirements.txt
python-dotenv==1.0.0
ccxt==4.1.13
pandas==2.1.4
numpy==1.26.2
requests==2.31.0
python-binance==1.0.19
TA-Lib==0.4.28
pytest==7.4.3
pytest-mock==3.12.0
pytest-cov==4.1.0
colorama==0.4.6 